Change GTM warning message and Change gtm warning message and organize the project
