+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Change GTM warning message
+- Add missing dependencies 
+- Rename `ProductSearchContextProvider` to `SearchContextProvider`.
 
 ## [1.9.5] - 2018-08-08
 ### Fixed
