+++ react/components/withDataLayer.js
@@ -1,5 +1,5 @@
-import React, { Component } from 'react'
-import PropTypes from 'prop-types'
+import React, { Component } from '../../../../Library/Caches/typescript/2.9/node_modules/@types/react'
+import PropTypes from '../../../../Library/Caches/typescript/2.9/node_modules/@types/prop-types'
 
 const { Consumer, Provider } = React.createContext({
   dataLayer: [],
