+++ react/package.json
@@ -4,8 +4,12 @@
     "test": "jest"
   },
   "dependencies": {
-    "@vtex/styleguide": "^3.0.2",
+    "react": "^16.2.0",
+    "react-apollo": "^2.0.4",
+    "react-dom": "^16.2.0",
+    "react-intl": "^2.4.0",
     "prop-types": "^15.5.10",
+    "ramda": "^0.25.0",
     "react-slick": "^0.16.0"
   },
   "devDependencies": {
@@ -19,14 +23,8 @@
     "enzyme-adapter-react-16": "^1.1.1",
     "enzyme-react-intl": "^1.4.8",
     "enzyme-to-json": "^3.3.1",
-    "graphql": "^0.13.0",
-    "graphql-tag": "^2.7.3",
     "jest": "^22.2.1",
     "jest-transform-graphql": "^2.1.0",
-    "react": "^16.2.0",
-    "react-apollo": "^2.0.4",
-    "react-dom": "^16.2.0",
-    "react-intl": "^2.4.0",
     "react-test-renderer": "^16.2.0"
   },
   "jest": {
